backend/requirements.txt
a2wsgi==1.10.10
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
docx2pdf==0.1.8
et_xmlfile==2.0.0
fastapi==0.135.3
google-auth==2.49.1
google-genai==1.70.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
kombu==5.6.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-http-client==3.3.7
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
ruff==0.15.9
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.43.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.8
